Controller.Host: extend map_device method
map_device method only had option to map based on
the MAC address of the device and as comment mentioned
it was not sufficient for public usage.

With SRIOV VF device having random MAC address in the
default namespace, we cannot rely on the MAC address
mapping, thus adding deterministing ifname mapping
is the way.